Delete all (#174)
* Implement `delete --all`


* linties


* Quick fix for arg length in delete

* Quick fix for arg length in delete

* Set min to zero

* Clarify error message

+++ cmd/ksync/delete.go
@@ -33,7 +33,7 @@ func (d *deleteCmd) new() *cobra.Command {
 		Long:    long,
 		Example: example,
 		Aliases: []string{"d"},
-		Args:    cobra.MinimumNArgs(1),
+		Args:    cobra.MinimumNArgs(0),
 		Run:     d.run,
 	})
 
@@ -52,11 +52,16 @@ func (d *deleteCmd) new() *cobra.Command {
 
 func (d *deleteCmd) run(cmd *cobra.Command, args []string) {
 	if viper.GetBool("all") {
-		d.deleteAll()
-	} else {
+		if len(cmd.Flags().Args()) == 0 {
+			d.deleteAll()
+		}
+		log.Fatal("cannot specify names when using `--all`")
+	} else if len(cmd.Flags().Args()) != 0 {
 		for name := range args {
 			d.delete(args[name])
 		}
+	} else {
+		log.Errorf("delete requires at least one spec name or `--all`")
 	}
 }
